chore(rust/chunk): prepare moving `Chunk` to `rolldown_common`
