+++ crates/rolldown/src/lib.rs
@@ -20,8 +20,8 @@ pub(crate) type SharedResolver = Arc<Resolver<OsFileSystem>>;
 pub(crate) type SharedOptions = Arc<NormalizedBundlerOptions>;
 
 pub use crate::{
-  bundler::Bundler, bundler_builder::BundlerBuilder, chunk::render_chunk::PreRenderedChunk,
-  types::bundle_output::BundleOutput,
+  bundler::Bundler, bundler_builder::BundlerBuilder,
+  chunk::types::pre_renderer_chunk::PreRenderedChunk, types::bundle_output::BundleOutput,
 };
 
 pub use rolldown_common::bundler_options::*;
