+++ crates/rolldown/src/utils/mod.rs
@@ -4,6 +4,7 @@ use rolldown_oxc_utils::{AstSnippet, OxcAst};
 
 use super::finalizer::{Finalizer, FinalizerContext};
 
+pub mod chunk;
 pub mod load_source;
 pub mod normalize_options;
 pub mod renamer;
